multi-translation prototype (french)
Searching for ways to translate WLED without making it a burden for the developers.
Translations are not always correct, or correctly placed.
This is a WIP, feel free to contact me if you want to help, I'm currently busy on another project.

- tools/translate.js change all {{}} (mustache-like) fields with their translation.
- You can change the language using my_config.h
- This is an interesting approach has it doesn't use any space for the translation
- The downside : for every language you need to create a different version (like on tasmota) but since there is a web installer, on the user side it could be transparent (detect browser language and choose correct firmware)

- In order to correctly translate everything in FX.h
JSON_mode_names / palette should be moved to a different file and  in a #ifdef should be translated in all language

- You can also add localisation\translate.js in the html code to test the translation, client side (but it won't work on translation made in javascript)